college-chatbot-backend/requirements.txt
fastapi
uvicorn
pandas
chromadb
langchain-groq
langchain-community
python-dotenv
pydantic